Modify Float6x32/BFloat6x32 to Float6x16/BFloat6x16 in Tensile (#18)
* Modify packing size to 16 instead of 32

* Missing to apply clang-format

+++ projects/hipblaslt/tensilelite/client/src/DataInitialization.cpp
@@ -760,11 +760,12 @@ namespace TensileLite
             HIP_CHECK_EXC(
                 hipMemcpy(dst,
                           src,
-                          multiplyElementSize(totalElements, DataTypeInfo::Get(descriptor.dataType()).elementSize),
+                          multiplyElementSize(totalElements,
+                                              DataTypeInfo::Get(descriptor.dataType()).elementSize),
                           kind));
             ptrdiff_t dPadding = totalElements - descriptor.totalAllocatedElements();
-            dPadding = multiplyElementSize(dPadding, descriptor.elementBytes());
-            void* dstOffset = (void*)((uint8_t*)dst + dPadding / 2);
+            dPadding           = multiplyElementSize(dPadding, descriptor.elementBytes());
+            void* dstOffset    = (void*)((uint8_t*)dst + dPadding / 2);
             TensileLite::hip::CopyTensorVoid(dstOffset, src, descriptor, kind);
             return dstOffset;
         }
@@ -782,9 +783,13 @@ namespace TensileLite
             const size_t    numElementsToCopy
                 = (customPadding == -1) ? descriptor.totalAllocatedElements()
                                         : (descriptor.totalAllocatedElements() + customPadding);
-            uint8_t* dstOffset = (uint8_t*)dst + multiplyElementSize(dPadding, descriptor.elementBytes());
+            uint8_t* dstOffset
+                = (uint8_t*)dst + multiplyElementSize(dPadding, descriptor.elementBytes());
             HIP_CHECK_EXC(
-                hipMemcpy(dstOffset, src, multiplyElementSize(numElementsToCopy, descriptor.elementBytes()), kind));
+                hipMemcpy(dstOffset,
+                          src,
+                          multiplyElementSize(numElementsToCopy, descriptor.elementBytes()),
+                          kind));
             return dstOffset;
         }
 
@@ -794,7 +799,8 @@ namespace TensileLite
                                size_t                  totalElements,
                                hipMemcpyKind           kind)
         {
-            HIP_CHECK_EXC(hipMemcpy(dst, src, multiplyElementSize(totalElements, descriptor.elementBytes()), kind));
+            HIP_CHECK_EXC(hipMemcpy(
+                dst, src, multiplyElementSize(totalElements, descriptor.elementBytes()), kind));
             return dst;
         }
 
@@ -978,8 +984,8 @@ namespace TensileLite
                             calculateKforSwizzling(dataType, MiK, MiKv, PackK);
                             numAllocatedElements = getSwizzledTensorNumAllocatedElements(
                                 problem.tensors()[i], MiM_N, MiK, PackK);
-                            numAllocatedBytes
-                                = multiplyElementSize(numAllocatedElements, rocisa::GetElementSize(dataType));
+                            numAllocatedBytes = multiplyElementSize(
+                                numAllocatedElements, rocisa::GetElementSize(dataType));
                         }
 
                         pristine.maxElements = std::max(pristine.maxElements, numAllocatedElements);
@@ -1212,8 +1218,8 @@ namespace TensileLite
                     else if(m_curBoundsCheck == BoundsCheckMode::GuardPageFront
                             || m_curBoundsCheck == BoundsCheckMode::GuardPageBack)
                     {
-                        float dataTypeSize = DataTypeInfo::Get(p->first).elementSize;
-                        size_t roundUpSize = divideElementSize(pageSize, dataTypeSize);
+                        float        dataTypeSize = DataTypeInfo::Get(p->first).elementSize;
+                        unsigned int roundUpSize  = divideElementSize(pageSize, dataTypeSize);
                         p->second.maxElements
                             = RoundUpToMultiple<size_t>(p->second.maxElements, roundUpSize);
                         // No bias page guard
@@ -1293,10 +1299,11 @@ namespace TensileLite
                     {
 
                         initArray(p.first, it.init, pUnit.cpuInput.valid.get(), pUnit.maxElements);
-                        HIP_CHECK_EXC(hipMemcpy(pUnit.gpuInput.valid.get(),
-                                                pUnit.cpuInput.valid.get(),
-                                                multiplyElementSize(pUnit.maxElements, dataTypeSize),
-                                                hipMemcpyHostToDevice));
+                        HIP_CHECK_EXC(
+                            hipMemcpy(pUnit.gpuInput.valid.get(),
+                                      pUnit.cpuInput.valid.get(),
+                                      multiplyElementSize(pUnit.maxElements, dataTypeSize),
+                                      hipMemcpyHostToDevice));
                     }
                     // Init and copy bad from cpu to gpu
                     if(pUnit.gpuInput.bad && pUnit.cpuInput.bad)
@@ -1305,10 +1312,11 @@ namespace TensileLite
                                   InitMode::BadOutput,
                                   pUnit.cpuInput.bad.get(),
                                   pUnit.maxElements);
-                        HIP_CHECK_EXC(hipMemcpy(pUnit.gpuInput.bad.get(),
-                                                pUnit.cpuInput.bad.get(),
-                                                multiplyElementSize(pUnit.maxElements, dataTypeSize),
-                                                hipMemcpyHostToDevice));
+                        HIP_CHECK_EXC(
+                            hipMemcpy(pUnit.gpuInput.bad.get(),
+                                      pUnit.cpuInput.bad.get(),
+                                      multiplyElementSize(pUnit.maxElements, dataTypeSize),
+                                      hipMemcpyHostToDevice));
                     }
                 }
             }
@@ -1321,7 +1329,8 @@ namespace TensileLite
                 for(auto& p : it.pristine)
                 {
                     auto&  pUnit = p.second;
-                    size_t size  = multiplyElementSize(pUnit.maxElements, DataTypeInfo::Get(p.first).elementSize);
+                    size_t size  = multiplyElementSize(pUnit.maxElements,
+                                                      DataTypeInfo::Get(p.first).elementSize);
                     if(size <= 0)
                     {
                         throw std::runtime_error("Size not exists.");
@@ -1393,7 +1402,8 @@ namespace TensileLite
                 for(auto& p : it.pristine)
                 {
                     auto&  pUnit = p.second;
-                    size_t size  = multiplyElementSize(pUnit.maxElements, DataTypeInfo::Get(p.first).elementSize);
+                    size_t size  = multiplyElementSize(pUnit.maxElements,
+                                                      DataTypeInfo::Get(p.first).elementSize);
 
                     std::stringstream ss;
                     ss << "[" << tensorIdx << "]" << "Failed to allocate gpu input " << it.name
@@ -1536,7 +1546,8 @@ namespace TensileLite
                                       problem.tensors()[i], MiM_N, MiK, PackK);
                     }
                 }
-                padding = multiplyElementSize(padding, DataTypeInfo::Get(problem.tensors()[i].dataType()).elementSize);
+                padding = multiplyElementSize(
+                    padding, DataTypeInfo::Get(problem.tensors()[i].dataType()).elementSize);
                 uint8_t* offset = (uint8_t*)pUnit.gpuInput.current.get();
                 initGPUBatchedInput((void*)(offset + padding),
                                     pUnit.gpuInput.batch.get(),
@@ -1560,9 +1571,11 @@ namespace TensileLite
                                   - problem.tensors()[ContractionProblemGemm::TENSOR::BIAS]
                                         .totalAllocatedElements();
                     }
-                    padding = multiplyElementSize(padding, DataTypeInfo::Get(
-                               problem.tensors()[ContractionProblemGemm::TENSOR::BIAS].dataType())
-                               .elementSize);
+                    padding = multiplyElementSize(
+                        padding,
+                        DataTypeInfo::Get(
+                            problem.tensors()[ContractionProblemGemm::TENSOR::BIAS].dataType())
+                            .elementSize);
                     uint8_t* offset = (uint8_t*)pUnitBias.gpuInput.current.get();
                     initGPUBatchedInput((void*)(offset + padding),
                                         pUnitBias.gpuInput.batch.get(),
@@ -1685,8 +1698,8 @@ namespace TensileLite
                                         problem.gemms[j].metadataLayout());
                                 }
                             }
-                            gemmInitOffset
-                                += multiplyElementSize(p.second.groupedGemmOffsets[j], tensors[i].elementBytes());
+                            gemmInitOffset += multiplyElementSize(p.second.groupedGemmOffsets[j],
+                                                                  tensors[i].elementBytes());
                         }
                     }
                 }
@@ -2246,61 +2259,63 @@ namespace TensileLite
                 setContractionInputs(u8Ptr, batchPtrs, ws, cdata, maxElements, isGPU, &unit);
                 inputs->grouped.push_back(unit);
 
-                u8Ptr[ContractionProblemGemm::TENSOR::A]
-                    += multiplyElementSize(offsets[ContractionProblemGemm::TENSOR::A][idx], problem.a().elementBytes());
-                u8Ptr[ContractionProblemGemm::TENSOR::B]
-                    += multiplyElementSize(offsets[ContractionProblemGemm::TENSOR::B][idx], problem.b().elementBytes());
-                u8Ptr[ContractionProblemGemm::TENSOR::C]
-                    += multiplyElementSize(offsets[ContractionProblemGemm::TENSOR::C][idx], problem.c().elementBytes());
-                u8Ptr[ContractionProblemGemm::TENSOR::D]
-                    += multiplyElementSize(offsets[ContractionProblemGemm::TENSOR::D][idx], problem.d().elementBytes());
+                u8Ptr[ContractionProblemGemm::TENSOR::A] += multiplyElementSize(
+                    offsets[ContractionProblemGemm::TENSOR::A][idx], problem.a().elementBytes());
+                u8Ptr[ContractionProblemGemm::TENSOR::B] += multiplyElementSize(
+                    offsets[ContractionProblemGemm::TENSOR::B][idx], problem.b().elementBytes());
+                u8Ptr[ContractionProblemGemm::TENSOR::C] += multiplyElementSize(
+                    offsets[ContractionProblemGemm::TENSOR::C][idx], problem.c().elementBytes());
+                u8Ptr[ContractionProblemGemm::TENSOR::D] += multiplyElementSize(
+                    offsets[ContractionProblemGemm::TENSOR::D][idx], problem.d().elementBytes());
                 if(u8Ptr[ContractionProblemGemm::TENSOR::E] != nullptr)
                 {
-                    u8Ptr[ContractionProblemGemm::TENSOR::E]
-                        += multiplyElementSize(offsets[ContractionProblemGemm::TENSOR::E][idx],
-                                               problem.tensors()[ContractionProblemGemm::TENSOR::E].elementBytes());
+                    u8Ptr[ContractionProblemGemm::TENSOR::E] += multiplyElementSize(
+                        offsets[ContractionProblemGemm::TENSOR::E][idx],
+                        problem.tensors()[ContractionProblemGemm::TENSOR::E].elementBytes());
                 }
                 if(u8Ptr[ContractionProblemGemm::TENSOR::BIAS] != nullptr)
                 {
-                    u8Ptr[ContractionProblemGemm::TENSOR::BIAS]
-                        += multiplyElementSize(offsets[ContractionProblemGemm::TENSOR::BIAS][idx],
-                                               problem.tensors()[ContractionProblemGemm::TENSOR::BIAS].elementBytes());
+                    u8Ptr[ContractionProblemGemm::TENSOR::BIAS] += multiplyElementSize(
+                        offsets[ContractionProblemGemm::TENSOR::BIAS][idx],
+                        problem.tensors()[ContractionProblemGemm::TENSOR::BIAS].elementBytes());
                 }
                 if(u8Ptr[ContractionProblemGemm::TENSOR::SCALEA] != nullptr)
                 {
-                    u8Ptr[ContractionProblemGemm::TENSOR::SCALEA]
-                        += multiplyElementSize(offsets[ContractionProblemGemm::TENSOR::SCALEA][idx],
-                                               problem.tensors()[ContractionProblemGemm::TENSOR::SCALEA].elementBytes());
+                    u8Ptr[ContractionProblemGemm::TENSOR::SCALEA] += multiplyElementSize(
+                        offsets[ContractionProblemGemm::TENSOR::SCALEA][idx],
+                        problem.tensors()[ContractionProblemGemm::TENSOR::SCALEA].elementBytes());
                 }
                 if(u8Ptr[ContractionProblemGemm::TENSOR::SCALEB] != nullptr)
                 {
-                    u8Ptr[ContractionProblemGemm::TENSOR::SCALEB]
-                        += multiplyElementSize(offsets[ContractionProblemGemm::TENSOR::SCALEB][idx],
-                                               problem.tensors()[ContractionProblemGemm::TENSOR::SCALEB].elementBytes());
+                    u8Ptr[ContractionProblemGemm::TENSOR::SCALEB] += multiplyElementSize(
+                        offsets[ContractionProblemGemm::TENSOR::SCALEB][idx],
+                        problem.tensors()[ContractionProblemGemm::TENSOR::SCALEB].elementBytes());
                 }
                 if(u8Ptr[ContractionProblemGemm::TENSOR::SCALEC] != nullptr)
                 {
-                    u8Ptr[ContractionProblemGemm::TENSOR::SCALEC]
-                        += multiplyElementSize(offsets[ContractionProblemGemm::TENSOR::SCALEC][idx],
-                                               problem.tensors()[ContractionProblemGemm::TENSOR::SCALEC].elementBytes());
+                    u8Ptr[ContractionProblemGemm::TENSOR::SCALEC] += multiplyElementSize(
+                        offsets[ContractionProblemGemm::TENSOR::SCALEC][idx],
+                        problem.tensors()[ContractionProblemGemm::TENSOR::SCALEC].elementBytes());
                 }
                 if(u8Ptr[ContractionProblemGemm::TENSOR::SCALED] != nullptr)
                 {
-                    u8Ptr[ContractionProblemGemm::TENSOR::SCALED]
-                        += multiplyElementSize(offsets[ContractionProblemGemm::TENSOR::SCALED][idx],
-                                               problem.tensors()[ContractionProblemGemm::TENSOR::SCALED].elementBytes());
+                    u8Ptr[ContractionProblemGemm::TENSOR::SCALED] += multiplyElementSize(
+                        offsets[ContractionProblemGemm::TENSOR::SCALED][idx],
+                        problem.tensors()[ContractionProblemGemm::TENSOR::SCALED].elementBytes());
                 }
                 if(u8Ptr[ContractionProblemGemm::TENSOR::SCALEALPHAVEC] != nullptr)
                 {
-                    u8Ptr[ContractionProblemGemm::TENSOR::SCALEALPHAVEC]
-                        += multiplyElementSize(offsets[ContractionProblemGemm::TENSOR::SCALEALPHAVEC][idx],
-                                               problem.tensors()[ContractionProblemGemm::TENSOR::SCALEALPHAVEC].elementBytes());
+                    u8Ptr[ContractionProblemGemm::TENSOR::SCALEALPHAVEC] += multiplyElementSize(
+                        offsets[ContractionProblemGemm::TENSOR::SCALEALPHAVEC][idx],
+                        problem.tensors()[ContractionProblemGemm::TENSOR::SCALEALPHAVEC]
+                            .elementBytes());
                 }
                 if(u8Ptr[ContractionProblemGemm::TENSOR::Synchronizer] != nullptr)
                 {
-                    u8Ptr[ContractionProblemGemm::TENSOR::Synchronizer]
-                        += multiplyElementSize(offsets[ContractionProblemGemm::TENSOR::Synchronizer][idx],
-                                               problem.tensors()[ContractionProblemGemm::TENSOR::Synchronizer].elementBytes());
+                    u8Ptr[ContractionProblemGemm::TENSOR::Synchronizer] += multiplyElementSize(
+                        offsets[ContractionProblemGemm::TENSOR::Synchronizer][idx],
+                        problem.tensors()[ContractionProblemGemm::TENSOR::Synchronizer]
+                            .elementBytes());
                 }
             }
         }
